refactor: pc버전 헤더 프로필 클릭 시 드롭다운 디자인, 구성 수정
